# Increase Cat Performance With BufReader