Iterate over a file using BufReader's bytes() implementation over Files. This leads to a huge performance increase. This makes it almost identical to the C version's performance.

### With BufReader (cat -v mv)
real: 727ms
user: 193ms
sys:   37ms

### Without BufReader (cat -v mv)
real: 1205ms
user:  240ms
sys:   467ms